requirements.txt
torch
pandas
pydantic==1.10.8
fastapi==0.95.2
uvicorn==0.22.0
transformers==4.29.2
huggingface_hub==0.14.1
google-cloud